Preserve the line break when the user presses enter
Closes #2.  This makes it so that when the user presses the enter key, we preserve the resulting line break, until the user manually moves the selection cursor or presses enter again.  The idea is that by pressing the enter key, the user has expressed a desire that the document contain the resulting line break.

We forget any explicit line breaks when the user moves the cursor because this seems like more straightforward, "local" behavior.  It would seem a little weird to me if the user pressed enter, then heavily edited a different part of the document, and then came back and we continued preserving that one enter press from a long time ago.

An alternative approach would be to preserve all line breaks inserted by pressing the enter key, not just the most recent one.  However, I don't see how to do this efficiently.  Presumably, we would need to keep track of the line breaks using View.add_regions and View.get_regions, but these methods (probably) take an amount of time proportional to the total number of saved line breaks.  This might be okay in practice, but it seems bad in theory, and WrapAsYouType needs to be fast since it operates on every keystroke.

No matter what, the user will lose any saved enter presses if he closes and reopens a document.  There's no foolproof way to preserve all of the user's line breaks.  The "wrap_as_you_type_paragraphs" setting is the only way to permanently preserve paragraph breaks that cannot be inferred from indentation.  Still, it seems valuable to make some effort to respect the user's wishes when he presses the enter key.