Repeatedly join backwards
This changes word wrap fixup to continue performing backwards joins (i.e. to move the current line's text to the end of the previous line) as long as we should - until it is no longer possible to fit the first word in the current line at the end of the previous line, or until the two lines are no longer in the same paragraph and section.

Previously, we would only perform up to one backwards join.  However, this could lead to weird behavior, such as several consecutive keystrokes resulting in one backwards join each.  Allowing for an unlimited number of backwards joins is analogous to the fact that we allow for an unlimited number of forwards joins.

Test plan:

Open a document with several consecutive short lines that are in a wrappable section.  Add a letter to the last line.  Make sure the line is joined with as many preceding lines as possible.