Don't wrap paragraphs with true "single_line" entries
This changes WrapAsYouType so that a line whose "wrap_as_you_type_paragraphs" element has a "single_line" entry of true is not wrapped, and may extend past the wrap width.  This behavior seems more proper, and it's useful in at least one case: Python doctest formatting.  Wrapping a doctest line would cause it to break.